requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
tortoise-orm==0.19.2
aerich==0.7.1